Moved some models to the model submodule
